security layer refactor
