+++ TestIt.API/appsettings.json
@@ -4,5 +4,10 @@
   },
   "ConnectionStrings": {
     "DefaultConnection": "Data Source=test-it.database.windows.net;Initial Catalog=DbTodo;user id=adm;password=root@123;"
+  },
+  "Authentication": {
+    "SecretKey": "mysupersecret_secretkey!123",
+    "Audience": "ExampleAudience",
+    "Issuer": "ExampleIssuer"
   }
 }
