+++ TestIt.Security/TokenProviderOptions.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using Microsoft.IdentityModel.Tokens;
+
+namespace TestIt.Security
+{
+    public class TokenProviderOptions
+    {
+        public string Path { get; set; } = "/token";
+
+        public string Issuer { get; set; }
+
+        public string Audience { get; set; }
+
+        public TimeSpan Expiration { get; set; } = TimeSpan.FromMinutes(5);
+
+        public SigningCredentials SigningCredentials { get; set; }
+    }
+}
