+++ TestIt.sln
@@ -5,13 +5,15 @@ VisualStudioVersion = 15.0.26403.7
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestIt.API", "TestIt.API\TestIt.API.csproj", "{9FC25BAF-8306-4954-A3E9-43D481E9704E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestIt.Utils", "TestIt.Security\TestIt.Utils.csproj", "{A6C4ED78-C398-4F74-8E5B-4ABBDFC2BA47}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestIt.Business", "TestIt.Business\TestIt.Business.csproj", "{70D7F369-B6CD-4A12-A6EA-75982FA12CD4}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestIt.Data", "TestIt.Data\TestIt.Data.csproj", "{9A2B1055-50CB-4568-B19C-A0635C1E043F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestIt.Model", "TestIt.Model\TestIt.Model.csproj", "{1AFF1B94-DD7E-4C50-B57F-E164D8B0A1B2}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestIt.Security", "TestIt.Security\TestIt.Security.csproj", "{106EBEC8-EF15-4C43-9D01-BB1DAD2006E4}"
+EndProject
+Project("{E53339B2-1760-4266-BCC7-CA923CBCF16C}") = "docker-compose", "docker-compose.dcproj", "{8C10F7C6-74A4-4FD6-9132-C5E5A64DBA1E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,10 +25,6 @@ Global
 		{9FC25BAF-8306-4954-A3E9-43D481E9704E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9FC25BAF-8306-4954-A3E9-43D481E9704E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9FC25BAF-8306-4954-A3E9-43D481E9704E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A6C4ED78-C398-4F74-8E5B-4ABBDFC2BA47}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A6C4ED78-C398-4F74-8E5B-4ABBDFC2BA47}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A6C4ED78-C398-4F74-8E5B-4ABBDFC2BA47}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A6C4ED78-C398-4F74-8E5B-4ABBDFC2BA47}.Release|Any CPU.Build.0 = Release|Any CPU
 		{70D7F369-B6CD-4A12-A6EA-75982FA12CD4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{70D7F369-B6CD-4A12-A6EA-75982FA12CD4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{70D7F369-B6CD-4A12-A6EA-75982FA12CD4}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -39,6 +37,14 @@ Global
 		{1AFF1B94-DD7E-4C50-B57F-E164D8B0A1B2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1AFF1B94-DD7E-4C50-B57F-E164D8B0A1B2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1AFF1B94-DD7E-4C50-B57F-E164D8B0A1B2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{106EBEC8-EF15-4C43-9D01-BB1DAD2006E4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{106EBEC8-EF15-4C43-9D01-BB1DAD2006E4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{106EBEC8-EF15-4C43-9D01-BB1DAD2006E4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{106EBEC8-EF15-4C43-9D01-BB1DAD2006E4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8C10F7C6-74A4-4FD6-9132-C5E5A64DBA1E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8C10F7C6-74A4-4FD6-9132-C5E5A64DBA1E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8C10F7C6-74A4-4FD6-9132-C5E5A64DBA1E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8C10F7C6-74A4-4FD6-9132-C5E5A64DBA1E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
